+++ CHANGELOG.md
@@ -7,13 +7,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 * Show a warning when a stage has too many layers.
+* Show a warning on duplicate ids
 
 ### Changed
 * Fixes inconsistent `layer.setSize()` method. Now it has same arguments as any container.
 * Full rewrite to Typescript with tons of refactoring and small optimizations. The public API should be 100% the same
 * Fixed `patternImage` and `radialGradient` for `Konva.Text`
 * `Konva.Util._isObject` is renamed to `Konva.Util._isPlainObject`.
-* TODO: changed behavior of `removeId`.
+* changed behavior of `removeId`.
 
 ### Removed
 * `Konva.Util.addMethods`